Reorganized tool and resource registration
