+++ packages/server-mcp/src/tools/mcp-tool-contribution.ts
@@ -1,0 +1,37 @@
+/********************************************************************************
+ * Copyright (c) 2025 EclipseSource and others.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License v. 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0.
+ *
+ * This Source Code may also be made available under the following Secondary
+ * Licenses when the conditions for such availability set forth in the Eclipse
+ * Public License v. 2.0 are satisfied: GNU General Public License, version 2
+ * with the GNU Classpath Exception which is available at
+ * https://www.gnu.org/software/classpath/license.html.
+ *
+ * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
+ ********************************************************************************/
+
+import { injectable, multiInject } from 'inversify';
+import { GLSPMcpServer, McpServerContribution, McpToolHandler } from '../server';
+
+/**
+ * MCP server contribution that provides tools for performing actions on
+ * GLSP diagrams, including validation and element creation.
+ *
+ * This contribution should not be overriden or extended if another resource is required.
+ * Instead, a new {@link McpToolHandler} should be registered like:
+ * @example
+ * bindAsService(bind, McpToolHandler, CreateNodeMcpToolHandler);
+ */
+@injectable()
+export class McpToolContribution implements McpServerContribution {
+    @multiInject(McpToolHandler)
+    protected mcpToolHandlers: McpToolHandler[];
+
+    configure(server: GLSPMcpServer): void {
+        this.mcpToolHandlers.forEach(handler => handler.registerTool(server));
+    }
+}
